+++ .env
@@ -1,0 +1,6 @@
+PORT = 3000
+DB_HOST = localhost
+DB_PORT = 3306
+DB_USER = root
+DB_PASSWORD = Ueba-2013
+DB_DATABASE = db_events
