+++ src/db.js
@@ -1,9 +1,15 @@
 import { createPool } from 'mysql2/promise'
-
+import {
+  DB_HOST,
+  DB_USER,
+  DB_PASSWORD,
+  DB_PORT,
+  DB_DATABASE
+} from './config.js'
 export const conn = createPool({
-  host: 'localhost',
-  user: 'root',
-  password: 'Ueba-2013',
-  port: '3306',
-  database: 'db_events'
+  host: DB_HOST,
+  user: DB_USER,
+  password: DB_PASSWORD,
+  port: DB_PORT,
+  database: DB_DATABASE
 })
