+++ src/index.js
@@ -1,20 +1,5 @@
-import express from "express";
-import userRoutes from './routes/user.routes.js'
-import indexRoutes from './routes/index.routes.js'
-import eventRoutes from './routes/event.routes.js'
+import app from "./app.js";
+import {PORT} from './config.js'
 
-const app = express();
-app.use(express.json());
-
-app.use('/user', userRoutes);
-app.use('/event', eventRoutes);
-app.use(indexRoutes);
-
-app.use((req, res, next) => {
-  res.status(404).json({
-    message: 'Endpoint not found'
-  })
-});
-
-app.listen(3000);
-console.log('Server 3000');
+app.listen(PORT);
+console.log('Server running on port ', PORT);
