+++ src/routes/index.routes.js
@@ -3,6 +3,6 @@ import { ping } from "../controllers/index.controller.js";
 
 const router = Router();
 
-router.get('/ping', ping);
+router.get('', ping);
 
 export default router;
